windows build is now working and initial windows ci setup

+++ rtl/ctypes.pas
@@ -10,7 +10,7 @@ type
     {$else}
         {$ifdef WINDOWS}
         cint = integer;
-        clong = int64;
+        clong = integer;
         {$endif}
     {$endif}
 implementation
